refactor: reorganize component structure and add currency management features
